svg/scripted/text-attrs-dxdy-have-length.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml"><metadata><h:link href="https://svgwg.org/svg2-draft/single-page.html#text-TSpanNotes" rel="help"/><h:meta content="dx and dy attributes on text elements are lists that support length()" name="assert"/></metadata><text id="text" x="120" y="150" fill="#00f" dx="60 90 -30 120 60 -257" dy="0 12 24 12 0 -12 -24 -12 0" font-family="Verdana" font-size="55">His socks are black.</text><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><script><![CDATA[
    test(function() {
      var text = document.getElementById('text');
      assert_equals(text.dx.baseVal.length, 6);
      assert_equals(text.dy.baseVal.length, 9);
    });
  ]]></script></svg>
